requirements.txt
ipwhois
python-whois
python-docx
PyYAML
webtech
xmltodict